@@ -5,24 +5,24 @@
         /// <summary>
         /// This sproc will be used when getting the instances.
         /// </summary>
-        string GetStoredProcedureName { get; }
+        string? GetStoredProcedureName { get; }
 
         /// <summary>
         /// This sproc will be used when putting (usually Table Type objects) 
         /// </summary>
-        string PutStoredProcedureName { get; }
+        string? PutStoredProcedureName { get; }
 
         /// <summary>
         /// Turn this flag on when the alternate sproc should be used instead of the StoredProcedureName
         /// </summary>
 
-        bool UseAlternateSproc { get; }
+        bool? UseAlternateSproc { get; }
 
         /// <summary>
         /// Is the sproc being used, uses table types instead
         /// of just raw parameters?
         /// </summary>
-        bool UsesTableTypes { get; }
+        bool? UsesTableTypes { get; }
 
     }
 }
