@@ -8,7 +8,7 @@ namespace Json.Orm
 {
     public partial class GenericStatusResponse
     {
-        public List<Response> Responses { get; set; }
+        public List<Response>? Responses { get; set; }
 
         public string ToJson() => JsonSerializer.Serialize(this);
 
@@ -18,11 +18,11 @@ namespace Json.Orm
     {
         public int? HttpStatus { get; set; }
 
-        public string Severity { get; set; }
+        public string? Severity { get; set; }
 
-        public string Message { get; set; }
+        public string? Message { get; set; }
 
-        public string Specific { get; set; }
+        public string? Specific { get; set; }
 
         public DateTimeOffset Utc { get; set; }
 
